# openapi/v1.yaml
...
openapi: 3.0.3
info:
  title: JumpWeb API
  version: 1.0.0
  description: |
    API de la plataforma de reservas JumpWeb. Sus clientes son la SPA del sidebar (Fase 4) y la
    app móvil (Fase 6); ambos son clientes de igual rango sobre el mismo dominio.

    **Autenticación** (spec §4.2): la SPA de primera parte usa la cookie de sesión de Sanctum con
    CSRF —nunca un token en `localStorage`—; el cliente móvil usa un Bearer emitido por
    `POST /auth/tokens`, con caducidad y revocación.

    **Errores**: todos comparten un único sobre (`Error`). El campo que un cliente debe programar
    es `code`, no `message` ni el texto de ninguna traducción.

    **Idioma**: se negocia por `Accept-Language` entre `es`, `en` y `fr`. Un cliente que comparta
    sesión con la web (la SPA) hereda la elección del visitante sin enviar nada.
servers:
  - url: /api/v1
    description: La versión va en la URL. `v1` es evolutiva hasta el primer cliente móvil (Fase 6).
...
paths:
...
  /auth/google/complete:
    post:
      tags: [Sesión]
      operationId: completeGoogleSignup
      summary: Completar el alta con Google
      description: |
        Crea la cuenta con lo que Google **no** puede dar —teléfono y aceptaciones— y **abre sesión**:
        quien llega acaba de demostrar su identidad ante Google y no tiene contraseña que teclear (es
        la misma decisión que el alta suelta desde `#331`).

        En la misma operación se escriben los sellos legales, una fila de `consents` por tipo, el
        vínculo con Google y **la firma del descargo**, que aquí se puede hacer en el acto porque el
        correo queda verificado (`[DECIDIDO owner]`).

        ⚠️ **Consume el perfil**: un segundo envío —dos pestañas, doble clic— ya no encuentra nada que
        crear y responde 404.

        ⚠️ **La privacidad no es una casilla** (§7.1): el RGPD pide INFORMAR, no que se acepte, y la
        base legal de una reserva es el contrato. El rastro sí se escribe.

        ⚠️ El **marketing no se pide aquí** (art. 7.4, no empaquetado): vive en el interruptor de la
        cuenta.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleSignupRequest'
      responses:
        '201':
          description: Cuenta creada y sesión abierta, **sin cuerpo**. El cliente pide `GET /me`.
        '400':
          description: |
            La petición no puede abrir sesión (no viene de un origen declarado *stateful*), y sin
            sesión no hay perfil de Google que consumir. No es falta de identidad: es que el cliente
            no habla el modo SPA.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No hay ningún alta de Google esperando (caducó, ya se usó, o nunca la hubo).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: |
            El texto del descargo se publicó de nuevo entre servirlo y aceptarlo
            (`waiver_document_stale`). El cliente vuelve a pedir `GET /legal/waiver`, lo presenta
            otra vez y **desmarca la casilla**: lo que se leyó ya no es lo que se firmaría.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/Maintenance'
...
  /legal/waiver:
    get:
      tags: [Instalación]
      operationId: getWaiverDocument
      summary: El texto firmable del waiver, vigente
      description: |
        El SNAPSHOT vigente del waiver en el idioma negociado, con el identificador que el cliente
        devuelve al aceptar (`POST /auth/register` con `accept_waiver`, o `POST /me/waiver`).
        Fase 6 (`specs/waiver-probatorio.md` §4.2, §4.4).

        **Público a propósito**: quien se está dando de alta todavía no tiene sesión y es justo cuando
        lo necesita. Lo que publica es lo que se firma —una versión inmutable, ya interpolada—, nunca
        la página editable del CMS.

        `document` es `null` cuando no hay nada que firmar aquí: `mode` distinto de `interno`, o ninguna
        versión publicada todavía. **No es un error**: es un estado legítimo de la instalación.

        Si el idioma pedido no se publicó en la versión vigente, cae al de respaldo (el mismo respaldo
        que aplica la web al pintar el texto) y `document.locale` dice cuál se sirvió: el idioma es
        parte de la prueba.
      responses:
        '200':
          description: El modo de la instalación y, si lo hay, el texto vigente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaiverDocument'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/Maintenance'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: laravel_session
      description: |
        Modo SPA de primera parte. El cliente pide primero `GET /sanctum/csrf-cookie` y envía el
        token CSRF en `X-XSRF-TOKEN`. Solo funciona desde un origen declarado como *stateful*.
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Modo cliente nativo. Token emitido por `POST /auth/tokens` (paso 3), con caducidad y
        revocable desde el borrado RGPD, el cambio de contraseña y «cerrar otras sesiones».
  responses:
...
    ValidationFailed:
      description: Los datos enviados no validan. `error.fields` los enumera por campo.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Limitador alcanzado. `params.retry_after` indica los segundos que faltan.
      headers:
        Retry-After:
          description: Segundos hasta que se puede reintentar.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Maintenance:
      description: El sitio está en mantenimiento (#218). Alcanza también a la API, a propósito.
      headers:
        Retry-After:
          description: Segundos orientativos hasta el próximo intento.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      description: |
        Sobre único de error de toda la API (spec §4.3). `code` es el contrato estable; `message`
        es texto ya traducido, para mostrar cuando el cliente no sabe hacer nada mejor.
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              description: |
                Código estable. La lista es CERRADA: añadir uno es evolutivo, renombrarlo o
                quitarlo es incompatible y exige `v2`. Debe coincidir con `App\Http\Api\ApiErrorCode`.
              enum:
                - unauthenticated
                - invalid_credentials
                - unauthorized
                - not_found
                - method_not_allowed
                - session_expired
                - validation_failed
                - too_many_requests
                - maintenance
                - bad_request
                - server_error
                # Negocio — la cesta no se puede convertir en pedido (todos con 422).
                - cart_empty
                - cart_too_large
                - product_unavailable
                - line_unavailable
                - line_past_date
                - line_too_late
                - line_too_soon
                - line_outside_window
                - line_sold_out
                - line_pack_sold_out
                - line_pack_guests_range
                - line_event_required
                # Negocio — la reserva no se admite (409) o el cobro no se abre (502).
                - reservations_paused
                - too_many_pending_orders
                - order_not_retryable
                - guest_form_closed
                - payment_unavailable
                # Fase 6 · waiver — aceptar el texto (409): no se firma aquí, o el texto cambió.
                - waiver_not_internal
                - waiver_document_stale
                - waiver_email_unverified
                # Fase 6 · menores a cargo (422 los dos).
                - dependent_not_minor
                - dependents_limit_reached
                # T5 · supresión (art. 17): con reservas por celebrar la cuenta no se borra (409).
                - account_has_upcoming_reservations
            message:
              type: string
              description: Mensaje legible en el idioma negociado.
            params:
              type: object
              description: |
                Datos para interpolar el mensaje en el cliente. Presente solo cuando aporta algo:
                hoy, `retry_after` en el 429; desde el paso 4, `product`/`when` de las reglas de
                reserva, sin los cuales el cliente pintaría «El producto — no está disponible».
              additionalProperties: true
            fields:
              type: object
              description: Errores por campo. Presente solo en `validation_failed`.
              additionalProperties:
                type: array
                items:
                  type: string
...
    GoogleSignupRequest:
      type: object
      description: |
        Lo único que Google no puede dar. **Ni el `sub` ni el correo viajan aquí** y no es un olvido:
        los pone el servidor desde la sesión (§6.3·3).
      required: [name, phone, accept_terms]
      additionalProperties: false
      properties:
        name:
          type: string
          maxLength: 255
        phone:
          type: string
          maxLength: 30
          description: Obligatorio — «imprescindible para las reservas» (`[DECIDIDO owner]`).
        accept_terms:
          type: boolean
          description: Aceptación contractual de las condiciones. Obligatoria.
        accept_waiver:
          type: boolean
          nullable: true
          description: |
            El descargo de responsabilidad. **Obligatorio cuando `GET /legal/waiver` sirve un
            documento** (modo interno con versión publicada); fuera de ese caso no se pide, y
            entonces esta pantalla ni se pinta.
        waiver_document_id:
          type: integer
          nullable: true
          minimum: 1
          description: |
            El identificador del texto que el servidor SIRVIÓ. Aceptar sin decir qué se leyó no prueba
            nada: si no es el vigente, la respuesta es `409 waiver_document_stale`.
...
    WaiverDocument:
      type: object
      description: El modo de la instalación y, si lo hay, el texto firmable vigente (Fase 6 · waiver).
      required: [mode, document]
      additionalProperties: false
      properties:
        mode:
          type: string
          enum: [externo, interno, desactivado]
          description: Cómo gestiona el waiver esta instalación (`waiver.mode`).
        document:
          type: object
          nullable: true
          description: '`null` fuera del modo interno o sin versión publicada. No es un error.'
          required: [id, version, locale, title, sections, published_at]
          additionalProperties: false
          properties:
            id:
              type: integer
              description: El identificador que se devuelve al aceptar. Cambia con cada versión publicada.
            version:
              type: integer
              description: Número de versión del texto (1, 2, …), común a todos sus idiomas.
            locale:
              type: string
              description: El idioma en que se sirvió — puede ser el de respaldo. Es parte de la prueba.
            title:
              type: string
            sections:
              type: array
              description: El texto íntegro, ya interpolado, como lista de secciones encabezado + párrafo.
              items:
                type: object
                required: [h, p]
                additionalProperties: false
                properties:
                  h:
                    type: string
                    description: Encabezado de la sección; puede ser cadena vacía.
                  p:
                    type: string
            published_at:
              type: string
              format: date-time
              nullable: true
...